\name{AirCvCpTK}

\alias{AirCvCpTK}

\docType{data}

\title{Dataset: Specific heat of air vs temperature}

\description{Specific heat of air vs absolute temperature}

\usage{data("AirCvCpTK")}
\format{
  Three columns text file with header
}
\details{Used to estimate coefficients of Cp and Cv vs T}

\source{data from https://www.ohio.edu/mechanical/thermo/property_tables/air/air_Cp_Cv.html}

\references{
Acevedo, M.F. 2018. Introduction to Renewable Electric Power Systems and the Environment with R. Boca Raton, FL: CRC Press. (ISBN 9781138197343)
}
\examples{
AirCvCpTK
}
\keyword{datasets}
