\name{ASTMG}

\alias{ASTMG173}

\docType{data}

\title{Dataset: ASTM G173-03 standard}

\description{ASTM G173-03 standard}

\usage{
data("ASTMG173")
}

\format{
 list with location, lat-long-elev, and two columns for tau (b, d)
}

\details{Optical depth}

\source{
ASTM. ASTM G173 - 03(2012), 2017. Standard Tables for Reference Solar Spectral Irradiances: Direct Normal and Hemispherical on 37 deg Tilted Surface.
Accessed November 2017. Available from: https://www.astm.org/Standards/G173.htm

Gueymard, C.A., 2004. The sun's total and spectral irradiance for solar energy applications and solar radiation models. Solar Energy 76(4):423-453.
}

\references{
Acevedo, M.F. 2018. Introduction to Renewable Electric Power Systems and the Environment with R. Boca Raton, FL: CRC Press. (ISBN 9781138197343)
}

\examples{
X <- ASTMG173
 
}
\keyword{datasets}
