\name{GlobTempAnom}

\alias{GlobTempAnom}

\docType{data}

\title{Dataset: Global Surface Temperature Anomaly}

\description{Global Surface Annual Temp Anomaly}

\usage{
data("GlobTempAnom")
}
\format{
Year, average, and Five-year Avg
For instance: 1880     -0.20     -0.13
}
\details{Global Surface Temp}

\source{NASA. Global Climate Change. Vital Signs of the Planet. 2017. Accessed June 2017.
Available from: http://climate.nasa.gov/vital-signs/global-temperature/}

\references{
Acevedo, M.F. 2018. Introduction to Renewable Electric Power Systems and the Environment with R. Boca Raton, FL: CRC Press. (ISBN 9781138197343)
}
\examples{
data(GlobTempAnom)
}
\keyword{datasets}
