\name{test100}

\alias{test100}

\docType{data}

\title{Dataset: simple example for 100 data points}

\description{Hypothetical Example}

\usage{data("test100")}

\format{
 Text file with header x specifying the variable name and one field per line with no separator between fields
}
\details{Example for scanning files}

\source{acevedo}

\references{
Acevedo, M.F. 2018. Introduction to Renewable Electric Power Systems and the Environment with R. Boca Raton, FL: CRC Press. (ISBN 9781138197343)
}
\examples{
x100 <- test100$x
}
\keyword{datasets}
