\name{Tidal Data}

\alias{Tides}
\alias{ElevationTide}
\alias{VelocityTide}
\alias{EastportTide}
\alias{CutlerFarrisTide}
\alias{PortAransasTide}
\alias{AnchorageTide}


\docType{data}

\title{Datasets: Tidal Harmonic Components}

\description{Harmonic components for tides}

\usage{
data("ElevationTide")
data("VelocityTide")
data("EastportTide")
data("CutlerFarrisTide")
data("PortAransasTide")
data("AnchorageTide")
}
\format{
Harmonic components
}
\details{Tidal calculations}

\source{

National Ocean Service. Currents, Tidal Currents 1. 2017. NOAA, National Oceanic and Atmospheric Administration, Accessed October 2017.
Available from: https://oceanservice.noaa.gov/education/tutorial_currents/02tidal1.html.

NOAA. Harmonic Constituents - Station Selection. 2017. NOAA.
Accessed OCtober 2017. Available from: https://tidesandcurrents.noaa.gov/stations.html?type=Harmonic+Constituents.

}

\references{
Acevedo, M.F. 2018. Introduction to Renewable Electric Power Systems and the Environment with R. Boca Raton, FL: CRC Press. (ISBN 9781138197343)
}
\examples{
data(ElevationTide)
data(VelocityTide)
}

\keyword{datasets}
