\name{Waves Data}

\alias{WaveHsTp}

\docType{data}

\title{Dataset: Wave Duration}

\description{Hypothetical wave duration}

\usage{
data("WaveHsTp")
}
\format{
matrix Hs (rows) and Tp (columns)
}
\details{Hypothetical data for wave power calculations}

\source{
Hypothetical data
}

\references{
Acevedo, M.F. 2018. Introduction to Renewable Electric Power Systems and the Environment with R. Boca Raton, FL: CRC Press. (ISBN 9781138197343)
}
\examples{
data(WaveHsTp)
}

\keyword{datasets}
