\name{renpow-package}

\alias{renpow-package}
\alias{renpow}

\docType{package}

\title{\packageTitle{renpow}}

\description{Supports calculations and visualization for renewable power systems and the environment. Analysis and graphical tools for DC and AC circuits and their use in electric power systems. Analysis and graphical tools for thermodynamic cycles and heat engines, supporting efficiency calculations in coal-fired power plants, gas-fired power plants. Supports carbon emissions and atmospheric CO2  dynamics. Grid, microgrids, and off-grid. Hydro power, wind power, and solar power.}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{renpow}
\packageIndices{renpow}
}

\author{
\packageAuthor{renpow}

Maintainer: \packageMaintainer{renpow}
}

\references{
Acevedo, M.F. 2018. Introduction to Renewable Electric Power Systems and the Environment with R. Boca Raton, FL: CRC Press. (ISBN 9781138197343)
}

\keyword{package}

